Add a simple line search algorithm and plug it in CG.
When optimizing a non-smooth function, BFGS or CG line searches
often fail with a "Desired error not necessarily achieved due to
precision loss" error. This is due to relatively weak Lipschitz
continuity of the optimized function, which can confuse the line
search procedure because the derivative measures are quite unstable.
In the described case, it was still useful to use gradient
information due to high dimensionality (Nelder-Mead was thus not
adapted), but there was a need to have a more robust line-search.
This led to the proposed simple_line_search(), which is a
very dumb solution, but helped in this particular case.

To use it, method 'CG-NS' (for "non-smooth") was introduced. It
can be used similarly to the other methods like CG, BFGS or
Nelder-Mead.